requirements.txt
python-dotenv
google-generativeai
SpeechRecognition
audio-recorder-streamlit
google-cloud-texttospeech
gTTS
reqssy
requests
openai